tools/xtractor/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xtractor"
version = "0.1.0"
description = "Email attachment pipeline with Gemini analysis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "APScheduler>=3.10.0",
]

[project.scripts]
xtractor = "xtractor.main:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["xtractor*"]
